requirements.txt
telebot
python-dotenv
requests
psycopg2-binary
matplotlib
apscheduler
pytz
